Move computed props logic into its own file.
