+++ client/js/app-state/index.js
@@ -1,10 +1,10 @@
+import compute from '../compute';
 import cookie from '../cookie';
 import { sortObjectsByDate, clearLocalStorage, getViewFromHref, applyFilters } from '../utils';
 
 export default function getInitialState () {
   let loggedIn = !!cookie.get('logged_in');
   if(!loggedIn) clearLocalStorage();
-  let _filterText = '';
   let entries = JSON.parse(localStorage.getItem('entries')) || undefined;
 
   const computedProps = {
@@ -15,10 +15,10 @@ export default function getInitialState () {
   };
 
   let state = {
-    entries: [],
     entryIndex: -1,
     filterText: '',
     viewEntries: [],
+    entries: [],
     entry: undefined,
     scrollPosition: 0,
     loggedIn: loggedIn,
@@ -28,27 +28,7 @@ export default function getInitialState () {
     dark: localStorage.getItem('dark') === 'true'
   };
 
-  let compute = {
-    set: function(obj, prop, value) {
-      obj[prop] = value;
-
-      for(let computedProp in computedProps){
-        let computedConfig = computedProps[computedProp];
-        if(~computedConfig.args.indexOf(prop)){
-          let computedArgs = computedConfig.args.map(arg => {
-            return obj[arg];
-          });
-          obj[computedProp] = computedConfig.computer(...computedArgs);
-        }
-      }
-
-      return true;
-    }
-  };
-
-  let stateProxy = new Proxy(state, compute);
-
+  let stateProxy = compute(state, computedProps);
   stateProxy.entries = entries;
-
   return stateProxy;
 };
